Feature/133313 importacao boletins

+++ src/SME.SGP.Dominio/Entidades/ProficienciaIdep.cs
@@ -3,10 +3,10 @@
     public class ProficienciaIdep: EntidadeBase
     {
         public string CodigoEOLEscola { get; set; }
-        public long SerieAno { get; set; }
+        public int SerieAno { get; set; }
         public int ComponenteCurricular { get; set; }
         public decimal Proficiencia { get; set; }
-        public long AnoLetivo { get; set; }
+        public int AnoLetivo { get; set; }
         public string Boletim { get; set; }
     }
 }
